athlete_os_env/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "fidenz-athlete-os"
version = "0.1.0"
description = "Multi-sport player simulation platform — OpenEnv-compliant RL environment"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Fidenz Labs"}]
keywords = ["openenv", "sports", "reinforcement-learning", "graphrag", "multi-agent"]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "numpy>=1.24.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "websockets>=12.0",
    "python-multipart>=0.0.6",
    "pyyaml>=6.0",
    "openenv-core>=0.2.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff", "black"]

[project.scripts]
server = "server.app:main"
